Enforce mobx actions and fix async action mutations not being wrapped
Ideally we would use mobx.flow for this but for some reason (probably a config issue) the flow code example throws TS2683 (implicit any type on this) for us (and suppressing the error with config change reveals that `this` isn't correctly referencing the store object). Using mobx.runInAction instead until this is solved.